refactor: split code into separate modules
