+++ src/utils/getPosition.ts
@@ -1,0 +1,8 @@
+import { mouse } from "@nut-tree/nut-js";
+
+const getPosition = async () => {
+  const position = await mouse.getPosition();
+  return [position.x, position.y];
+};
+
+export default getPosition;
